Combining CSV
I have never used Github before and this is new to me so if I am doing something wrong please let me know. I will be rewatching the professor's videos on Github in the meanwhile. I am trying to combine the data but I keep getting an error. I think it's because the files are too large for my laptop, but I am unsure. If someone could try this code and let me know if it works for them that would be great.