Update Combine_csv
Rosa's code worked fine for me. There is only a very small error I corrected in this code. "nflId" for the last join had the D capitalized in the original  file and was giving me an error. It ran fine once this was changed. The dataset is really large (70 million+ observations) so it runs slow but it does work.